"collection" and "item" usage
Add a section on the IANA-registered "collection" and "item"
link relations.  Describe the common create-via-collection
pattern.  While colleciton resources described with these
link relations are not required to implement such creation
semantics, they are forbidden from implementing incompatible
behavior when used with protocols such as HTTP where the pattern
is well-established.

Note, of course, that extended link relation types may be defined
to identify collections with non-standard behavior in a Hyper-Schema
context.