Fix bugs in draft-04 schema.json meta-schema
"format" was omitted, and "uri" was mis-used, as both "id"
and "$schema" can accept URI-references.

While "regex" was accidentally removed from draft-04 as a
format, since that was a bug and not an intentional change, and
since unrecognized formats are to be ignored, this leaves the
"regex" format in place in the meta-schema for "pattern".