Split "Metadata" in to re-use + annotations
In preparation for the applicability text, split out "definitions"
into its own section about re-use, and rename the remainder of
the old metadata section to focus on annotation of schemas with
keywords or new vocabularies.

Rework the "definitoins" section text, but leave the other
sections alone until the next commit.