Add recursiveRoot and recursiveRef
If the root schema of a schema document contains

`"$recursiveRoot": true`

then that root schema is set for that schema document
and any to which it referes as the target of all
`"$recursiveRef"` references, regardless of their values.

Encountering further `"$recursiveRoot"` keywords in
root schemas of referenced documents does **not** further
change the target.  This will be explained in detail
with comments added to the diff.

`"$recursiveRoot"` MUST be ignored in subschemas.

If no `"$recursiveRoot": true"` has been encountered,
then `"$recursiveRef"` is evaluated exactly as if it
were `"$ref"`.  Its value is a URI reference, which is
resolved according to the usual rules involving `"$id".

The key point is that someone reading the schema will
know that a `"$recursiveRef"` might have its target
changed dynamically at runtime, while `"$ref"` never will.

-----------

The following changes were made:

* schema.json, hyper-schema.json, and the additional example of
  hyper-operations.json (further extending hyper-schema.json)
  all have `"$recursiveRoot": true`
* links.json does not use `"$recursiveRoot"`
* The reference from each extension meta-schema to its "base"
  is a _normal_ `"$ref"`.  Otherwise it would be an infinite loop.
* All other schema references become `"$recursiveRef"`, with the
  same URI Reference value as before
* All of the properties and $defs duplicated from schema.json
  to hyper-schema.json can now be removed

Note that there were several odd trailing "#" fragments, which
should not be present in `"$id"` in particular, so I dropped those.
They are not part of this change I just found them surprising.